[1/3] omap4xxx: DOMX: Disable Google DOMX build in non enhanced mode
1/3 part patch for omap4 devices to use custom "domx" source via
2 BoardConfig.mk items:
ENHANCED_DOMX and DOMX_PATH.  ENHANCED_DOMX is used as a check
for optionally stopping the domx source in omap4xxx and also
pointing to openmax includes via DOMX_PATH in the other 2 parts
of this patch.

The new source for domx libs would be located in the device tree.

This commit for hardware/ti/omap4xxx is a cherry-pick from omapzoom:
http://www.omapzoom.org/?p=platform/hardware/ti/omap4xxx.git;a=commitdiff;h=040534d546dbcfcff16ff1afff15c01c2822dfcc

Change-Id: I676bf8722f16584022a7888c0ff3fe6c7f9096f0
Signed-off-by: Hashcode <hashcode0f@gmail.com>